src/requirements.txt
Django>=1.11.5
psycopg2
requests
djangorestframework
markdown
django-filter
django-cors-headers
djangorestframework-camel-case
django-taggit
django-taggit-serializer
six
gunicorn
dj-database-url
bootstrap-admin
django-excel
pyexcel-xls
pyexcel-xlsx
pyexcel-ods
gunicorn